[DOCKER-388] Configure alfresco repository to not download POM files
ASS 2.0.2 has a parent pom that points to a parent pom that is in a
private repository.

`org.alfresco:alfresco-search-services:2.0.2` ->
`org.alfresco:alfresco-search-parent:2.0.2` ->
`org.alfresco:alfresco-search-and-insight-parent:2.0.2` (private)

Because we don't need any dependencies from the POMs anyways, we can
configure Gradle to not download any metadata and just take the
artifacts directly.
Because we don't want this behavior for all artifacts, the repository is
limited to the relevant modules.